media-player: Rework styling for media player
Fix the styling for the media player and make it look the same in the
notification drawer and on the lockscreen.

Turns the media player widget into a GtkListBoxRow and places it
in a GtkListBox so that the styling is inherited from libhandy.
This way the same style can also be applied to the notifications
without having the css duplicated.
Fixes: https://source.puri.sm/Librem5/phosh/-/issues/371
Fixes: https://source.puri.sm/Librem5/phosh/-/issues/370