PhoshHome: Reset overview when folding ends instead of when unfolding starts
Resetting the app grid from filtered state can take a noticeable
amount of time. Do it when folding animation ends, when it's
far less likely that the user wants to interact with the shell,
rather than when unfolding animation starts, which is almost
guaranteed to be done because of user action. Thanks to that
the shell can respond faster.